rename gxff package to xff so it's not strictly for grpc
