+++ http_server.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/packethost/hegel/gxff"
+	"github.com/packethost/hegel/xff"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
@@ -23,8 +23,8 @@ func ServeHTTP() {
 		logger.Fatal(err, "could not register custom endpoints")
 	}
 
-	trustedProxies := gxff.ParseTrustedProxies()
-	http.Handle("/", handleTrustedProxies(mux, trustedProxies))
+	trustedProxies := xff.ParseTrustedProxies()
+	http.Handle("/", xff.HTTPHandler(logger, mux, trustedProxies))
 
 	logger.With("port", *metricsPort).Info("Starting http server")
 	go func() {
